add some required packages in the virtualenv
this python interpreter is used by the runbot, whereas buildout
environment uses some tweaking of sys.path to work. Without this all the builds
are failing due to missing dependencies.

requirements.txt is also the place to list things which are external
dependencies of the addons we want to test.